Split blockV1 and blockV2
